# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server",
    "description": "",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.0.1"
  },
  "paths": {
    "/api/artists": {
      "get": {
        "tags": [
          "artists"
        ],
        "operationId": "list_artists",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number, 1-indexed. Defaults to 1.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page. Defaults to 20. The server enforces a maximum.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Optional text search filter.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of artists ordered by name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_ArtistSummary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "artists"
        ],
        "operationId": "create_artist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArtistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created artist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/artists/{id}": {
      "get": {
        "tags": [
          "artists"
        ],
        "operationId": "get_artist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artist detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "artists"
        ],
        "operationId": "update_artist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArtistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated artist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "artists"
        ],
        "operationId": "delete_artist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/artists/{id}/images": {
      "post": {
        "tags": [
          "artists"
        ],
        "operationId": "upload_artist_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImageUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Image uploaded and linked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistImageInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Artist not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/artists/{id}/images/{image_id}": {
      "delete": {
        "tags": [
          "artists"
        ],
        "operationId": "delete_artist_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "image_id",
            "in": "path",
            "description": "Image ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistImageInfo": {
        "type": "object",
        "description": "An image attached to an artist with its semantic role.",
        "required": [
          "id",
          "public_url",
          "kind"
        ],
        "properties": {
          "credits": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "public_url": {
            "type": "string",
            "description": "Publicly served URL for clients."
          }
        }
      },
...
      "ArtistLinkInfo": {
        "type": "object",
        "description": "An external link associated with an artist.",
        "required": [
          "id",
          "url",
          "kind"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": "string"
          }
        }
      },
      "ArtistLinkInput": {
        "type": "object",
        "description": "Input for creating an external link on an artist.",
        "required": [
          "url",
          "kind"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ArtistLinkKind"
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": "string"
          }
        }
      },
      "ArtistLinkKind": {
        "type": "string",
        "description": "Valid platform kinds for an artist external link.",
        "enum": [
          "youtube",
          "website",
          "x",
          "instagram",
          "twitch",
          "other"
        ]
      },
      "ArtistResponse": {
        "type": "object",
        "description": "A full artist record returned by detail and mutation endpoints.",
        "required": [
          "id",
          "name",
          "images",
          "links"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistImageInfo"
            }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistLinkInfo"
            }
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "CreateArtistRequest": {
        "type": "object",
        "description": "Request body for `POST /api/artists`.",
        "required": [
          "name",
          "links"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistLinkInput"
            }
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Body returned for all error responses.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "ImageUpload": {
        "type": "object",
        "description": "Placeholder schema for image multipart upload bodies.",
        "required": [
          "file",
          "kind"
        ],
        "properties": {
          "credits": {
            "type": [
              "string",
              "null"
            ]
          },
          "file": {
            "type": "string",
            "format": "binary"
          },
          "kind": {
            "type": "string",
            "description": "Semantic role of the image. See [`ArtistImageKind`](api_types::artists::ArtistImageKind)."
          }
        }
      },
...
      "PagedResponse_ArtistSummary": {
        "type": "object",
        "description": "Generic envelope returned by all paginated list endpoints.",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "A lean artist record returned by the list endpoint.",
              "required": [
                "id",
                "name",
                "images"
              ],
              "properties": {
                "description": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "images": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtistImageInfo"
                  }
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "description": "The page that was returned, 1-indexed.",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int32",
            "description": "Items per page as applied by the server (may be less than requested if capped).",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of matching items across all pages.",
            "minimum": 0
          }
        }
      },
...
      "UpdateArtistRequest": {
        "type": "object",
        "description": "Request body for `PUT /api/artists/{id}`.\n\nImages are managed via the `/images` subresource.",
        "required": [
          "name",
          "links"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistLinkInput"
            }
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}